Possibly done with Manage

+++ src/chocAn/Manage.java
@@ -54,6 +54,48 @@ public class Manage {
 		System.out.println("In addMember()");
 		String name, number, address, city, state, zipCode, status;
 		//dialogue box required to get info
+		while(true){
+			name=(String)JOptionPane.showInputDialog(frame,"What is the member's name?\nName must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"John");
+			if(name.length()<=25){
+				break;
+			}
+		}
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the member's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		while(true){
+			address=(String)JOptionPane.showInputDialog(frame,"What is the member's address?\nAddress must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"30 2nd St");
+			if(address.length()<=25){
+				break;
+			}
+		}
+		while(true){
+			city=(String)JOptionPane.showInputDialog(frame,"What is the member's place of residence?\nCity name must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"Gotham");
+			if(city.length()<=14){
+				break;
+			}
+		}
+		while(true){
+			state=(String)JOptionPane.showInputDialog(frame,"What is the member's state of residence?\nState must use the 2 letter abbreviation","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"IN");
+			if(state.length()==2){
+				char first, second;
+				first=state.charAt(0);
+				second=state.charAt(1);
+				if(Character.isUpperCase(first)&&Character.isUpperCase(second)){
+					break;
+				}
+			}
+		}
+		while(true){
+			zipCode=(String)JOptionPane.showInputDialog(frame,"What is the member's zipCode??\nZipCode must contain 5 integers","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"37188");
+			if(zipCode.length()==5){
+				break;
+			}
+		}
+		
 		Object[] possibilites = {"Active","Nonactive"};
 		status = (String)JOptionPane.showInputDialog(frame, "What is the member's status?","Add Member",JOptionPane.PLAIN_MESSAGE, icon, possibilites ,"Active");
 		
@@ -84,6 +126,13 @@ public class Manage {
 		System.out.println("In updateMember()");
 		String name, number, address, city, state, zipCode, status;
 		//dialogue box required to get info
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the member's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		
 		Member temp = new Member();
 		Member curr = temp.getMemberByNumber(number);
 		
@@ -92,6 +141,51 @@ public class Manage {
 		}else{
 			String oldnum = curr.getNumber();
 			//dialogue box to add new information
+			while(true){
+				name=(String)JOptionPane.showInputDialog(frame,"What is the member's name?\nName must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"John");
+				if(name.length()<=25){
+					break;
+				}
+			}
+			while(true){
+				number=(String)JOptionPane.showInputDialog(frame,"What is the member's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+				if(number.length()==9){
+					break;
+				}
+			}
+			while(true){
+				address=(String)JOptionPane.showInputDialog(frame,"What is the member's address?\nAddress must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"30 2nd St");
+				if(address.length()<=25){
+					break;
+				}
+			}
+			while(true){
+				city=(String)JOptionPane.showInputDialog(frame,"What is the member's place of residence?\nCity name must be under 26 characters","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"Gotham");
+				if(city.length()<=14){
+					break;
+				}
+			}
+			while(true){
+				state=(String)JOptionPane.showInputDialog(frame,"What is the member's state of residence?\nState must use the 2 letter abbreviation","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"IN");
+				if(state.length()==2){
+					char first, second;
+					first=state.charAt(0);
+					second=state.charAt(1);
+					if(Character.isUpperCase(first)&&Character.isUpperCase(second)){
+						break;
+					}
+				}
+			}
+			while(true){
+				zipCode=(String)JOptionPane.showInputDialog(frame,"What is the member's zipCode??\nZipCode must contain 5 integers","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"37188");
+				if(zipCode.length()==5){
+					break;
+				}
+			}
+			
+			Object[] possibilites = {"Active","Nonactive"};
+			status = (String)JOptionPane.showInputDialog(frame, "What is the member's status?","Add Member",JOptionPane.PLAIN_MESSAGE, icon, possibilites ,"Active");
+			
 			curr.setName(name);
 			curr.setNumber(number);
 			curr.setAddress(address);
@@ -112,6 +206,13 @@ public class Manage {
 		System.out.println("In deleteMember()");
 		String number;
 		//dialogue box required to get info
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the member's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Member",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		
 		Member temp = new Member();
 		Member curr = temp.getMemberByNumber(number);
 		//code to search list && set boolean
@@ -130,7 +231,7 @@ public class Manage {
 		String operation, ret;
 		JFrame frame = new JFrame("ChocAn");
 		ImageIcon icon = new ImageIcon("choclate.png");
-		Object[] possibilities = {"Add Provider", "Update Provider", "Delete Provider"};
+		Object[] possibilities = {"Add Provider", "Update Provider", "Delete Provider","Exit"};
 		//Object[] possibilities2 = {"Card","Number"};
 		operation = (String)JOptionPane.showInputDialog(frame, "Which operation would you like to perform?","Manage Provider",JOptionPane.PLAIN_MESSAGE, icon, possibilities, "Add Provider");
 		
@@ -138,8 +239,10 @@ public class Manage {
 			ret=addProvider();
 		}else if (operation == "Update Provider"){
 			ret=updateProvider();
-		}else{
+		}else if(operation=="Delete Provider"){
 			ret=deleteProvider();
+		}else{
+			ret="Exiting manage provider";
 		}
 		JOptionPane.showMessageDialog(frame, ret);
 	}
@@ -151,8 +254,49 @@ public class Manage {
 	public String addProvider(){
 		System.out.println("In addProvider()");
 		String name, number, address, city, state, zipCode;
-		boolean found = false;
 		//dialogue box required to get info
+		while(true){
+			name=(String)JOptionPane.showInputDialog(frame,"What is the provider's name?\nName must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"CocoINC");
+			if(name.length()<=25){
+				break;
+			}
+		}
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the provider's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		while(true){
+			address=(String)JOptionPane.showInputDialog(frame,"What is the provider's address?\nAddress must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"30 2nd St");
+			if(address.length()<=25){
+				break;
+			}
+		}
+		while(true){
+			city=(String)JOptionPane.showInputDialog(frame,"What is the provider's place of residence?\nCity name must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"Gotham");
+			if(city.length()<=14){
+				break;
+			}
+		}
+		while(true){
+			state=(String)JOptionPane.showInputDialog(frame,"What is the provider's state of residence?\nState must use the 2 letter abbreviation","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"IN");
+			if(state.length()==2){
+				char first, second;
+				first=state.charAt(0);
+				second=state.charAt(1);
+				if(Character.isUpperCase(first)&&Character.isUpperCase(second)){
+					break;
+				}
+			}
+		}
+		while(true){
+			zipCode=(String)JOptionPane.showInputDialog(frame,"What is the provider's zipCode??\nZipCode must contain 5 integers","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"37188");
+			if(zipCode.length()==5){
+				break;
+			}
+		}
+		
 		Provider temp = new Provider();
 		Provider poss = temp.getProviderByNumber(number);
 		//code to search list && set boolean
@@ -169,7 +313,6 @@ public class Manage {
 			curr.addProviderToDatabase();
 			return "The provider has been added to the ChocAn database.";
 		}
-		return "Task failed";
 	}
 	
 	/**
@@ -179,25 +322,71 @@ public class Manage {
 	public String updateProvider(){
 		System.out.println("In updateProvider()");
 		String name, number, address, city, state, zipCode;
-		boolean found = false;
 		//dialogue box required to get info
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the provider's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		
 		Provider temp = new Provider();
 		Provider poss = temp.getProviderByNumber(number);
 		//code to search list && set boolean
 		if (poss==null){
 			return "This provider does not exist in the ChocAn database.";
 		}else{
-			Provider curr = new Provider();
-			curr.setName(name);
-			curr.setNumber(number);
-			curr.setAddress(address);
-			curr.setCity(city);
-			curr.setState(state);
-			curr.setZipCode(zipCode);
-			curr.addProviderToDatabase();
+			String oldnum=poss.getNumber();
+			while(true){
+				name=(String)JOptionPane.showInputDialog(frame,"What is the provider's name?\nName must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"CocoINC");
+				if(name.length()<=25){
+					break;
+				}
+			}
+			while(true){
+				number=(String)JOptionPane.showInputDialog(frame,"What is the provider's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+				if(number.length()==9){
+					break;
+				}
+			}
+			while(true){
+				address=(String)JOptionPane.showInputDialog(frame,"What is the provider's address?\nAddress must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"30 2nd St");
+				if(address.length()<=25){
+					break;
+				}
+			}
+			while(true){
+				city=(String)JOptionPane.showInputDialog(frame,"What is the provider's place of residence?\nCity name must be under 26 characters","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"Gotham");
+				if(city.length()<=14){
+					break;
+				}
+			}
+			while(true){
+				state=(String)JOptionPane.showInputDialog(frame,"What is the provider's state of residence?\nState must use the 2 letter abbreviation","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"IN");
+				if(state.length()==2){
+					char first, second;
+					first=state.charAt(0);
+					second=state.charAt(1);
+					if(Character.isUpperCase(first)&&Character.isUpperCase(second)){
+						break;
+					}
+				}
+			}
+			while(true){
+				zipCode=(String)JOptionPane.showInputDialog(frame,"What is the provider's zipCode??\nZipCode must contain 5 integers","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"37188");
+				if(zipCode.length()==5){
+					break;
+				}
+			}
+			poss.setName(name);
+			poss.setNumber(number);
+			poss.setAddress(address);
+			poss.setCity(city);
+			poss.setState(state);
+			poss.setZipCode(zipCode);
+			poss.updateProviderInDatabase(oldnum);
 			return "The provider has been updated in the ChocAn database.";
 		}
-		return "Task failed";
 	}
 	
 	/**
@@ -206,18 +395,24 @@ public class Manage {
 	 */
 	public String deleteProvider(){
 		System.out.println("In deleteProvider()");
-		boolean found = false;
+		String number;
 		//dialogue box required to get info
+		while(true){
+			number=(String)JOptionPane.showInputDialog(frame,"What is the provider's number?\nNumber must be 9 integers\nInclude leading zeros as needed","Add Provider",JOptionPane.PLAIN_MESSAGE,icon,null,"123456789");
+			if(number.length()==9){
+				break;
+			}
+		}
+		
 		Provider temp = new Provider();
 		Provider poss = temp.getProviderByNumber(number);
 		//code to search list && set boolean
 		if (poss!=null){
-			
+			poss.deleteProviderFromDatabase();
 			return "This provider has been deleted from the ChocAn database.";
 		}else{
 			return "The provider does not exist in the ChocAn database.";
 		}
-		return "Task failed";
 	}
 
 }
